requirements.txt
datasets
Pillow
tqdm
flask
flask-cors
openai
python-dotenv
